Only send reload signal when necessary (#47)
Currently, we send multiple reload signals to the NATS server when we
detect any file changes within a watched directory. This causes
unnecessary logging and reloading.

This change only sends a signal if a file we're interested in has
changed. It ignores other file changes within the same directory. In
addition, we no longer log for every event, of which there can be many,
not all of them useful.