Add resource cleanup loop (#56)
With finalizers gone, the Kubernetes Informers no longer send us an
event when a resource is deleted. The resource just gets deleted. This
is good because this avoids potential cluster lockups. However, this
means we're in charge of resource cleanup now. Currently, streams and
consumers are not deleted from JetStream when a user deletes a Stream or
Consumer CRD.

This change adds a cleanup loop that periodically checks for deleted
CRDs and deletes their corresponding JetStream stream or consumer.