feat: creates consumers if not present in NATS
The jetstream-controller its an operate that should ensure that a system
is in a given state. During our tests we found out that if a consumers
or stream are deleted via code or through `nats-box` they are not
created again by the controller.

From my understanding this is happening because we are checking if the
observed generation number and generation number match in the CRD, which
on this case (a manually delete consumer) will not change and thus not
trigger a recreation.

The following PR aims to resources (consumers and streams) being created
in NATS in case they are not found by the client but there is a CRD for it.