feat: Prevent deletes in consumers and streams
Adding a `preventDelete` property for streams and consumers CRDS which
when set to true nack should be able to delete it, otherwise it should
just ignore that stream or consumer.

The use case for this is in a multi-cloud context when we have deploy
deployed in different clouds it becomes troublesome. For example, nack
on GKE might check that a stream CRD is missing and proceeds to delete
it, however that CRD exists in AWS and AKS and it should not be deleted
because its going to impact other applications.