Bump GH Action versions (Go; node12 deprecations)
Bump various GitHub Actions versions to be more current, to avoid the
node.js 12 deprecation notices.

Bump Go to 1.19 where applicable, and quote the version strings.

Note that actions/setup-go@v4 is now available, with more caching, but I
deliberately chose to not explore that at this time and instead to stick to the
v3 upgrade, which we've done in many other repos and know is safe.

Add a warning comment by the use of the bake-action.